Switch from yaml file to docblocks
