--- docs/options.yaml
@@ -1,50 +1,0 @@
----
-print-width: |
-  > **For readability we recommend against using more than 80 characters:**
-  >
-  > In code styleguides, maximum line length rules are often set to 100 or 120.
-  > However, when humans write code, they don't strive to reach the maximum
-  > number of columns on every line. Developers often use whitespace to break
-  > up long lines for readability. In practice, the average line length often
-  > ends up well below the maximum.
-  >
-  > Prettier, on the other hand, strives to fit the most code into every line.
-  > With the print width set to 120, prettier may produce overly compact, or
-  > otherwise undesirable code.
-
-single-quote: |
-  Notes:
-  * Quotes in JSX will always be double and ignore this setting.
-  * If the number of quotes outweighs the other quote, the quote which is less used will be used to format the string - Example: `"I'm double quoted"` results in `"I'm double quoted"` and `"This \"example\" is single quoted"` results in `'This "example" is single quoted'`.
-
-parser: |
-  Both the `babylon` and `flow` parsers support the same set of JavaScript
-  features (including Flow). Prettier automatically infers the parser from the
-  input file path, so you shouldn't have to change this setting.
-
-filepath: |
-  For example, the following will use `postcss` parser:
-
-  ```bash
-  cat foo | prettier --stdin-filepath foo.css
-  ```
-
-require-pragma: |
-  Prettier can restrict itself to only format files that contain a special comment, called a pragma, at the top of the file. This is very useful
-  when gradually transitioning large, unformatted codebases to prettier.
-
-  For example, a file with the following as its first comment will be formatted when `--require-pragma` is supplied:
-
-  ```js
-  /**
-   * @prettier
-   */
-  ```
-
-  or
-
-  ```js
-  /**
-   * @format
-   */
-  ```
