+++ package.json
@@ -55,7 +55,6 @@
     "eslint-plugin-react": "7.1.0",
     "humanize-string": "^1.0.1",
     "jest": "21.1.0",
-    "js-yaml": "^3.10.0",
     "markdown-magic": "^0.1.19",
     "mkdirp": "^0.5.1",
     "prettier": "1.7.0",
